LVM resize: Rename VG to avoid problems on Fedora systems.
Typical Fedora systems use VGs with the name "fedora".
This is true for workstations as well as official Fedora
images (e.g. Fedora 28 Server image for AArch64).

So when deactivating the VG "fedora" an error will occur, because
the current machine's rootfs/swap VG can't be deactivated.

To address this issue, this patch looks up the VG UUID and
renames the VG to "image" before resizing.
After resizing we undo the renaming.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>