Add OLPC XO-1.75
With a recent enough firmware and the 5.0 kernel, it runs Fedora well
enough. The computer uses OpenFirmware, not U-Boot.

Upon boot, the firmware looks for a \boot\olpc.fth on an USB stick, SD
card, and the internal eMMC, checks that it starts with a backslash and
executes it. Our boot script checks that the Devicetree is good enough
and the boots the first entry found in extlinux.conf.