Fix label parsing in OLPC boot scripts
For simplicity's sake, the extlinux.conf parser doesn't present a menu.
In an event of trouble, the user could just break into the OFW FORTH shell.
What it's supposed to do is to just parse the first entry, and not
proceed to read further if it encounters another label.

When a label is encountered, a check is done to see if "boot-device" variable
is already set. This does not work as intended -- a bit-wise "and" is done
between the variable address and length, as opposed to a logical one.
Bummer. Let's just check the size against zero and drop the address check.
The address is constant and always non-zero anyway.