sleep after partprobe
I often see wrong partion names detected in the script immediatly following
partprobe. This is when running on a slow old computer. I have not observed
this when I add a sleep after partprobe. Sleeping for 5 seconds might be
overkill but we do sleep 5 elsewhere in the script so I went with that here
too.