Update AllWinner 64 bit devices
The AllWinner 64 bit devices originally had a SPL (sunxi-spl.bin) and
a separate U-Boot (u-boot.itb) which we wrote out in two steps. There
is now a combined option with both the U-Boot FIT and SPL combined
into a single file like the 32 bit AllWinner devices so lets use that.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>