Add support for resizing and mounting Btrfs filesystems
As desktop variants of Fedora are now using Btrfs for the root
filesystem, we need to be able to grow the Btrfs filesystem when
the user requests to expand the filesystem to fill the card.

Additionally, the installer script needs to be able to manipulate
the contents of the root filesystem as needed to configure it for
the target board.

Reference: https://fedoraproject.org/wiki/Changes/BtrfsByDefault

Signed-off-by: Neal Gompa <ngompa13@gmail.com>